requirements.txt
feedparser>=6.0.0
anthropic>=0.40.0
python-dotenv>=1.1.0
python-telegram-bot>=21.0
requests>=2.31.0
supabase>=2.0.0
